# ToDoAspCoreExample/src/ToDoAspCoreExample/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "Sample todo-api",
    "title" : "Todo",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:3000"
  } ],
  "paths" : {
    "/todo/{user}/{id}/resolve" : {
      "put" : {
        "operationId" : "resolveTodo",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "user",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Todo"
                  },
                  "type" : "array"
                }
              }
            }
          }
        },
        "summary" : "resolveTodo",
        "tags" : [ "TodoService" ]
      }
    },
    "/todo/{user}/{id}" : {
      "delete" : {
        "operationId" : "removeTodo",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "user",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Todo"
                  },
                  "type" : "array"
                }
              }
            }
          }
        },
        "summary" : "removeTodo",
        "tags" : [ "TodoService" ]
      }
    },
    "/todo/{user}" : {
      "delete" : {
        "operationId" : "removeAllTodos",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "user",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "204" : { }
        },
        "summary" : "removeAll",
        "tags" : [ "TodoService" ]
      },
      "get" : {
        "operationId" : "listTodos",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "user",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Todo"
                  },
                  "type" : "array"
                }
              }
            }
          }
        },
        "summary" : "listTodos",
        "tags" : [ "TodoService" ]
      },
      "post" : {
        "operationId" : "addTodo",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "user",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Todo"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Todo"
                  },
                  "type" : "array"
                }
              }
            }
          }
        },
        "summary" : "addTodo",
        "tags" : [ "TodoService" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "Todo" : {
        "example" : {
          "id" : "id",
          "title" : "title",
          "when" : "2000-01-23T04:56:07.000+00:00",
          "resolved" : true
        },
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "resolved" : {
            "type" : "boolean"
          },
          "when" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "id" ],
        "title" : "Todo",
        "type" : "object"
      }
    }
  }
}